@@ -20,26 +20,27 @@
   },
   "license": "GPL-3.0-or-later",
   "author": "WartaPoirier Corp.",
-  "main": "dist/index.html",
   "scripts": {
     "build": "parcel build src/index.html",
     "dev": "parcel src/index.html"
   },
   "devDependencies": {
-    "@types/react": "^17.0.14",
-    "@types/react-dom": "^17.0.9",
-    "@types/react-redux": "^7.1.18",
-    "parcel-bundler": "^1.9.4",
+    "@parcel/transformer-sass": "^2.6.2",
+    "@types/react": "^18.0.15",
+    "@types/react-dom": "^18.0.6",
+    "@types/react-redux": "^7.1.24",
+    "parcel": "^2.6.2",
+    "parcel-bundler": "^1.3.1",
     "parcel-plugin-static-files-copy": "^2.6.0",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "react-hook-size": "^1.3.5",
-    "react-redux": "^7.2.4",
+    "react-redux": "^8.0.2",
     "react-tooltip": "^4.2.21",
-    "redux": "^4.1.0",
+    "redux": "^4.2.0",
     "redux-localstorage": "^0.4.1",
-    "sass": "^1.35.2",
-    "typescript": "^4.3.5"
+    "sass": "^1.53.0",
+    "typescript": "^4.7.4"
   },
   "staticFiles": {
     "staticPath": "assets",
